+++ src/cli.ts
@@ -5,6 +5,7 @@ import {
   ArgumentParser,
   ArgumentDefaultsHelpFormatter,
   Namespace,
+  SUPPRESS,
 } from "argparse";
 
 import { launch } from "./browser.js";
@@ -30,8 +31,8 @@ parser.add_argument("-a", "--args", {
 });
 parser.add_argument("-b", "--browser", {
   choices: Object.values(BrowserType),
-  default: defaultArgs.browser,
-  help: "Which browser family to use for this test.",
+  dest: "browser_prev",
+  help: SUPPRESS,
 });
 parser.add_argument("-d", "--user-data-dir", {
   help:
@@ -106,10 +107,8 @@ parser.add_argument("-t", "--timeout", {
   type: "int",
 });
 parser.add_argument("-u", "--url", {
-  help:
-    "The URL to run measurements against. Should be a full URL (i.e., " +
-    "at least a scheme and a domain).",
-  required: true,
+  dest: "url_prev",
+  help: SUPPRESS,
   type: URL,
 });
 parser.add_argument("-v", "--version", {
@@ -137,6 +136,19 @@ parser.add_argument("--width", {
   default: defaultArgs.viewport?.width,
   help: "The width of the browser viewport to use when loading pages.",
   type: "int",
+});
+parser.add_argument("url", {
+  help:
+    "The URL to run measurements against. Should be a full URL (i.e., " +
+    "at least a scheme and a domain)",
+  type: URL,
+});
+parser.add_argument("browser", {
+  default: defaultArgs.browser,
+  help:
+    "Which browser family to use for this test, one of: " +
+    Object.values(BrowserType).join(", "),
+  nargs: "?",
 });
 
 try {